backend/requirements.txt
# Core dependencies
flask==2.0.1
Werkzeug>=2.0.0
python-magic>=0.4.24
PyPDF2>=2.0.0
python-docx>=0.8.11
gunicorn==20.1.0
reportlab>=3.6.0
beautifulsoup4==4.12.2

# WebSocket dependencies with fixed versions
simple-websocket==0.10.1
flask-sock==0.5.2
greenlet==3.0.3  # Updated to match playwright requirement
eventlet==0.33.3

# NLP and ML dependencies
spacy>=3.5.0
scikit-learn>=1.0.0
numpy>=1.20.0

# File handling
flask-dropzone>=1.6.0

# Additional dependencies
flask-cors==3.0.10
python-dotenv==0.19.0

